generate iban

+++ src/main/java/org/example/service/AccountServiceImpl.java
@@ -2,12 +2,15 @@ package org.example.service;
 
 import org.example.entity.Account;
 import org.example.entity.Transaction;
+import org.example.model.dto.AccountBalanceInfoDto;
 import org.example.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityExistsException;
+import javax.persistence.EntityNotFoundException;
+import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
@@ -25,14 +28,14 @@ public class AccountServiceImpl implements AccountService {
     @Override
     public Account getByIban(String iban) {
         return accountRepository.findById(iban)
-                .orElseThrow(() -> new IllegalArgumentException(String.format("Account with IBAN %s not found", iban)));
+                .orElseThrow(() -> new EntityNotFoundException(String.format("Account with IBAN %s not found", iban)));
     }
 
     @Override
     public List<Account> searchByClient(String firstName, String lastName) {
         List<Account> accounts = accountRepository.search(firstName, lastName);
         if (accounts.isEmpty()) {
-            throw new IllegalArgumentException("No account found");
+            throw new EntityNotFoundException("No account found");
         }
         return accounts;
     }
@@ -41,7 +44,7 @@ public class AccountServiceImpl implements AccountService {
     public Account create(Account account) {
         Account accountEntity = getByIban(account.getIban());
         if (accountEntity != null) {
-            throw new IllegalArgumentException("This account already exists in system");
+            throw new EntityExistsException("This account already exists in system");
         }
         return accountRepository.save(account);
     }
@@ -52,19 +55,22 @@ public class AccountServiceImpl implements AccountService {
         List<Transaction> transactions = new ArrayList<>();
         transactions.addAll(account.getDebitTransactions());
         transactions.addAll(account.getCreditTransactions());
-        Collections.sort(transactions, Comparator.comparing(Transaction::getCreatedAt));
+        transactions.sort(Comparator.comparing(Transaction::getCreatedAt));
         return transactions;
     }
 
     @Override
-    public String retrievingAccountBalance(String iban) {
+    public AccountBalanceInfoDto retrievingAccountBalance(String iban) {
         Account account = getByIban(iban);
-        String balance = account.getBalance() + " " + account.getCurrencyCode();
-        return balance;
+        return new AccountBalanceInfoDto(account.getBalance().doubleValue(),account.getCurrencyCode());
     }
 
     @Override
     public void deleteAccountByIban(String iban) {
+        Account account = getByIban(iban);
+        if(!account.getBalance().equals(BigDecimal.valueOf(0))) {
+            throw new UnsupportedOperationException("Deletion is not possible cause account balance is positive");
+        }
         accountRepository.delete(getByIban(iban));
     }
 }
